backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pandas==2.2.3
openpyxl==3.1.5
SQLAlchemy==2.0.35
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-jose[cryptography]